refactor: 리뷰 작성 테스트를 `ReviewTest`로 이동
